// src/components/jsonjoy-builder/types/json-schema.ts
const simpleTypes = [
  'string',
  'number',
  'integer',
  'boolean',
  'object',
  'array',
  'null',
] as const

// Define base schema first - Zod is the source of truth
/** @public */
export const baseSchema = z.object({
  // Base schema properties
  $id: z.string().optional(),
  $schema: z.string().optional(),
  $ref: z.string().optional(),
  $anchor: z.string().optional(),
  $dynamicRef: z.string().optional(),
  $dynamicAnchor: z.string().optional(),
  $vocabulary: z.record(z.string(), z.boolean()).optional(),
  $comment: z.string().optional(),
  title: z.string().optional(),
  description: z.string().optional(),
  default: z.unknown().optional(),
  deprecated: z.boolean().optional(),
  readOnly: z.boolean().optional(),
  writeOnly: z.boolean().optional(),
  examples: z.array(z.unknown()).optional(),
  type: z.union([z.enum(simpleTypes), z.array(z.enum(simpleTypes))]).optional(),

  // String validations
  minLength: z.number().int().min(0).optional(),
  maxLength: z.number().int().min(0).optional(),
  pattern: z.string().optional(),
  format: z.string().optional(),
  contentMediaType: z.string().optional(),
  contentEncoding: z.string().optional(),

  // Number validations
  multipleOf: z.number().positive().optional(),
  minimum: z.number().optional(),
  maximum: z.number().optional(),
  exclusiveMinimum: z.number().optional(),
  exclusiveMaximum: z.number().optional(),

  // Array validations
  minItems: z.number().int().min(0).optional(),
  maxItems: z.number().int().min(0).optional(),
  uniqueItems: z.boolean().optional(),
  minContains: z.number().int().min(0).optional(),
  maxContains: z.number().int().min(0).optional(),

  // Object validations
  required: z.array(z.string()).optional(),
  minProperties: z.number().int().min(0).optional(),
  maxProperties: z.number().int().min(0).optional(),
  dependentRequired: z.record(z.string(), z.array(z.string())).optional(),

  // Value validations
  const: z.unknown().optional(),
  enum: z.array(z.unknown()).optional(),
})

// Define recursive schema type
/** @public */
export type JSONSchema =
  | boolean
  | (z.infer<typeof baseSchema> & {
      // Recursive properties
      $defs?: Record<string, JSONSchema>
      contentSchema?: JSONSchema
      items?: JSONSchema
      prefixItems?: JSONSchema[]
      contains?: JSONSchema
      unevaluatedItems?: JSONSchema
      properties?: Record<string, JSONSchema>
      patternProperties?: Record<string, JSONSchema>
      additionalProperties?: JSONSchema | boolean
      propertyNames?: JSONSchema
      dependentSchemas?: Record<string, JSONSchema>
      unevaluatedProperties?: JSONSchema
      allOf?: JSONSchema[]
      anyOf?: JSONSchema[]
      oneOf?: JSONSchema[]
      not?: JSONSchema
      if?: JSONSchema
      then?: JSONSchema
      else?: JSONSchema
    })
// ...
export type ObjectJSONSchema = Exclude<JSONSchema, boolean>